% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean_cache.R
\name{clean_fre_cache}
\alias{clean_fre_cache}
\title{Cleans the local cache folder}
\usage{
clean_fre_cache(cache_folder = get_default_cache_dir())
}
\arguments{
\item{cache_folder}{Path of cache folder to delete (default = get_default_cache_dir())}
}
\value{
A boolean indicating whether the operation succeeded (invisibly)
}
\description{
This function deletes all downloaded ZIP files, temporary files, and cached
datasets stored in the cache folder.
}
\examples{
\dontrun{
clean_fre_cache()
}
}
