% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fre_data.R
\name{download_read_fre_zip_files}
\alias{download_read_fre_zip_files}
\title{Downloads and reads FRE file from ftp}
\usage{
download_read_fre_zip_files(df_file_in, cache_folder)
}
\arguments{
\item{df_file_in}{A dataframe with information about file}

\item{cache_folder}{Path of cache folder to keep cache and zip files}
}
\value{
A list with all tables from zip file
}
\description{
Downloads and reads FRE file from ftp
}
\examples{
\dontrun{
# no example fits here
}
}
