% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_fre_zip_file.R
\name{gdfpd.read.zip.file.type.fre}
\alias{gdfpd.read.zip.file.type.fre}
\title{Reads folder for zip file post 2011 (internal)}
\usage{
gdfpd.read.zip.file.type.fre(rnd_folder_name, folder_to_unzip = tempdir())
}
\arguments{
\item{rnd_folder_name}{Folder name where unzipped files are available}

\item{folder_to_unzip}{Folder to unzip files}
}
\value{
A list with financial statements
}
\description{
Reads folder for zip file post 2011 (internal)
}
\examples{
# no example (this functions not used directly)
}
