% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fre_data2.R
\name{get_fre_data2}
\alias{get_fre_data2}
\title{Downloads and reads FRE datasets}
\usage{
get_fre_data2(
  first_year = lubridate::year(lubridate::today()) - 1,
  last_year = lubridate::year(lubridate::today()),
  cache_folder = get_default_cache_dir()
)
}
\arguments{
\item{first_year}{First year of selected data}

\item{last_year}{Last year of selected data}

\item{cache_folder}{Path of cache folder to keep cache and zip files (default = 'gfred_cache')}
}
\value{
A list of tibbles, where each element is a different table from the FRE system
}
\description{
The FRE (formularios de referencia) is the reporting system of companies
traded at B3. This function will access the CVM ftp and parse all available files according to user
choices such as dates and companies.
}
\examples{
\dontrun{
l_fre <- get_fre_data2(cache_folder = fs::file_temp("fre-cache"))
}
}
