% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fre_links.R
\name{get_fre_links}
\alias{get_fre_links}
\title{Fetches ftp contents and parse all available links}
\usage{
get_fre_links(
  companies_cvm_codes,
  first_year = 2010,
  last_year = lubridate::year(Sys.Date()),
  cache_folder
)
}
\arguments{
\item{companies_cvm_codes}{Numeric CVM code  of companies. IF set to NULL (default), will return data for all available companies.}

\item{first_year}{First year of selected data}

\item{last_year}{Last year of selected data}

\item{cache_folder}{Path of cache folder to keep cache and zip files}
}
\value{
A dataframe with links and files
}
\description{
Fetches ftp contents and parse all available links
}
\examples{
\dontrun{ # keep cran fast
df_ftp <- get_contents_ftp('http://dados.cvm.gov.br/dados/CIA_ABERTA/DOC/FRE/DADOS/')
}
}
