% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_fre_zip_file.R
\name{my_fix_cols}
\alias{my_fix_cols}
\title{Manipulates list from FRE zip file}
\usage{
my_fix_cols(df_in, df_info_file)
}
\arguments{
\item{df_in}{Entry dataframe}

\item{df_info_file}{Informations about file}
}
\value{
An organized dataframe
}
\description{
This function will add several columns to each table in fre list.
}
\examples{
\dontrun{
# no example here
}
}
