% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{my_merge_dfs_lists}
\alias{my_merge_dfs_lists}
\title{Merges (row wise) dataframes from different list, using names of dataframes as index}
\usage{
my_merge_dfs_lists(l_1, l_2)
}
\arguments{
\item{l_1}{First dataframe}

\item{l_2}{Second dataframe}
}
\value{
A list with row-joined dataframes (same names as l.1)
}
\description{
Merges (row wise) dataframes from different list, using names of dataframes as index
}
\examples{

l_1 <- list(x = data.frame(runif(10)) )
l_2 <- list(x = data.frame(runif(10)) )

l <- my_merge_dfs_lists(l_1, l_2)

}
