% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_fre_zip_file.R
\name{read_single_fre_zip_file}
\alias{read_single_fre_zip_file}
\title{Reads a single FRE zip file downloaded from Bovespa}
\usage{
read_single_fre_zip_file(
  my_zip_file,
  df_file_in,
  cache_folder = "gfred_cache",
  folder_to_unzip = tempdir()
)
}
\arguments{
\item{my_zip_file}{Full path to zip file}

\item{df_file_in}{Dataframe with information about file}

\item{cache_folder}{Path of cache folder to keep cache and zip files}

\item{folder_to_unzip}{Folder to unzip files (default = tempdir())}
}
\value{
A list with several dataframes containing financial statements
}
\description{
Reads a single FRE zip file downloaded from Bovespa
}
\examples{

\dontrun{
# no examples fit here
}

}
