% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_xml.R
\name{xml.fct.auditing}
\alias{xml.fct.auditing}
\title{Reads XML data for auditing}
\usage{
xml.fct.auditing(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for auditing
}
\examples{

# No example (INTERNAL)
}
