% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_xml.R
\name{xml.fct.stockholder}
\alias{xml.fct.stockholder}
\title{Reads XML data for stockholder data}
\usage{
xml.fct.stockholder(x)
}
\arguments{
\item{x}{A list with stockholder data}
}
\value{
A dataframe
}
\description{
Reads XML data for stockholder data
}
\examples{

# No example (INTERNAL)
}
